iommu/omap: check for valid archdata in attach_dev
A device is tied to an iommu through its archdata field, any device
requiring to be attached to an iommu_domain must have valid archdata
containing the necessary iommu information. This archdata information
to be added is SoC-specific. Add a check in the omap_iommu_attach_dev
to make sure that the device has non-NULL archdata before accessing
different SoC-specific fields of the archdata. This prevents a NULL
pointer dereference on any misconfigured devices.

Signed-off-by: Suman Anna <s-anna@ti.com>