ARM: dts: dra7-evm: Add timers to IPUs and DSPs
The BIOS System Tick timers have been added for the IPU and DSP
remoteproc devices for the DRA7 EVM board. The following timers
are chosen, as per the timers used on the current firmware images:
	IPU2: GPTimer 3
	IPU1: GPTimer 11
	DSP1: GPTimer 5
	DSP2: GPTimer 6

The timers are optional, but are mandatory to support device
management features such as power management and watchdog support.
The above are added to successfully boot and execute firmware images
configured with the respective timers, images that use internal
processor subsystem timers are not affected. The timers can be
changed or removed as per the system integration needs, if needed.

Each of the IPUs has two Cortex-M4 processors, and is currently
expected to be running in SMP-mode, so only a single timer suffices
to provide the BIOS tick timer. An additional timer should be added
for the second processor in IPU if it were to be run in non-SMP mode.
The timer value also needs to be unique from the ones used by other
processors so that they can be run simultaneously.

Signed-off-by: Suman Anna <s-anna@ti.com>