ARM: OMAP2+: remoteproc: add dra7 dsp boot control functions
DRA7xx has upto two instances of DSP processors, and each DSP needs
its boot vector to be programmed in a Control module register. The
Control module API is currently limited only to the machine layers,
and the OMAP remoteproc driver programs the DSP Boot vectors using
a specific platform ops function.

The format of this boot control register is different from earlier
OMAP2+ SoCs, and is programmed by the dra7_dsp_ctrl_write_boot_addr()
function. Add the proper pdata ops implementation wrappers over this
function for properly writing the boot vector for each of the DSPs
in DRA7xx, so that they can be supplied to the OMAP remoteproc driver
through the respective platform data and pdata quirks.

Signed-off-by: Suman Anna <s-anna@ti.com>