ARM: OMAP2+: Add pdata-quirks for DSP2 rproc on DRA74x
The pdata quirks for the DSP2 remote processor device, present
usually on DRA74x SoCs, has been added. The DSP2 remote processor
is another identical instance as the DSP1 remote processor. The
quirks are almost the same as the DSP1 remote processor on DRA7xx,
except for the boot vector programming wrapper function.

The pdata quirks will be removed once the dependencies against
the arch/arm layres are resolved.

Signed-off-by: Suman Anna <s-anna@ti.com>