ARM: dts: DRA72x: Add aliases for rproc nodes
Add aliases for all the 3 remote processor nodes common to
all DRA72x boards. The aliases uses the stem "rproc", and
are defined in the order of the most common processors on
the DRA72x family. The ids are same as DRA74x except for
the missing DSP2.

The aliases can be overridden, if needed, in the respective
derivative board dts files.

Signed-off-by: Suman Anna <s-anna@ti.com>