ARM: dts: beagle-x15: Enable the system mailboxes 5 and 6
Enable the System Mailboxes 5 and 6 and the corresponding child
sub-mailbox (legacy) nodes, in order to communicate with the
respective remote processors IPU1, IPU2, DSP1 and DSP2 from the
host. The Dual-Cortex M4 IPU1 and IPU2 processor sub-systems are
assumed to be running in SMP-mode, and hence needs only a single
sub-mailbox node each.

Signed-off-by: Suman Anna <s-anna@ti.com>