ARM: dts: DRA72x: Add legacy sub-mailbox nodes for IPU1, IPU2 & DSP1
Add the legacy sub-mailbox nodes that were used to communicate between
MPU and the remote processors IPU1, IPU2 and DSP1. These match the
respective node definitions on DRA74x to maintain compatiblity for
the equivalent remote processors. There is no DSP2 on DRA72x, and so
the corresponding sub-mailbox node is not added.

These sub-mailbox nodes were added to provide backward compatibility
to the mailbox configuration used on previous TI product kernels and
IPC product.

All these sub-mailbox nodes are left in disabled state, and should
be enabled (and modified if needed) as per the individual product
configuration in the corresponding board dts files.

Signed-off-by: Suman Anna <s-anna@ti.com>