ARM: dts: beagle-x15: Add mailboxes, timers for rprocs
Add the required mailboxes and system timers to all the remote
processor device nodes (IPU1, IPU2, DSP1 and DSP2) for the AM57xx
Beagle X15 board. The mailbox and timer data is identical to that
of the equivalent processors on the DRA7 EVM board to maintain
firmware compatibility between both the board variants.

The watchdog timer data is also added, but only for the IPU2
remote processor node. The current firmware images do not have
watchdog configured for other processors.

The properties are chosen based on current firmwares, but can be
changed or removed as per the system integration needs after
making the appropriate changes on the firmware side.

Signed-off-by: Suman Anna <s-anna@ti.com>