OMAPDSS: add hdmi ops to dra7-evm-encoder-tpd12s015
A version of this patch was done upstream for tpd12s015 driver, in
ti-linux-3.14 there is an extra driver called
dra7-evm-encoder-tpd12s015.c that handles some of the "idiosyncrasies"
of the dra7-evm hdmi design and it also needed the same set of changes.

Add pass-through functions for set_infoframe and set_hdmi_mode to
dra7-evm-encoder-tpd12s015 drivers.

Signed-off-by: Darren Etheridge <detheridge@ti.com>