remoteproc: wkup_m3_rproc: Introduce wkup_m3_copy_aux_data
Introduce wkup_m3_copy_aux_data to give users of wkup_m3_rproc the
ability to copy data to a predefined data region that can be accessed by
the firmware running on the CM3.

Aux Data region is defined as the last 255 bytes of the DMEM used by the
wkup_m3.

Useful for providing i2c sequences to the CM3 firmware for scaling PMIC
voltages during low power modes.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>