ARM: DRA7: hwmod_data: Add reset data to second MMU on DSPs
Each of the DSP processor subsystems in DRA7xx has two MMUs, one
for the processor port and another for an EDMA port. Both these
MMUs share a common reset line, the MMU on the EDMA port is
expected to be mirror-programmed alongside the primary MMU, so
the reset info and pdata quirks for reset management was added
at present only for the MMU associated with the processor port.

The omap_hwmod layer tries to initialize all the hwmods and put
them in idle state during bootup. This involves turning ON the
corresponding power domain and then idling the IP. For hwmods
with reset data, the idling is skipped, and the IP will be managed
by the corresponding driver. The lack of the reset data on the
second MMU though causes the corresponding processor subsystem
domain to be left in ON state, and any suspend/resume operation
with PM enabled fails to transition this domain (even when the
driver is absent or when the corresponding device nodes are
disabled) to the desired state.

Add the reset data to the second MMUs on both the DSPs so that
the power domain can transition successfully to the desired OFF
state on any PM suspend/resume operations. The pdata quirks
were not to be added though as the OMAP IOMMU driver releases
the reset once and programs both the MMUs together.

Signed-off-by: Suman Anna <s-anna@ti.com>