drivers: net: cpsw: switch-config: unknown vlan handling
Currently only known VLANs are handled inside the switch, when an
unknown VLAN is received the switch will drop the packets. This
ioctl allows user to subscribe the ports for unknown VLAN ids, then
switch will start forwarding unknown VLAN packets as configured by
the user. Additional features of the unknown vlan handling of the
switch other than port forwarding are untag on egress, registered
and unregistered multicast of the unknown VLAN ids

Signed-off-by: Mugunthan V N <mugunthanvnm@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>