OMAPDSS: PLL: fix PLL config when pixel clock is low
The VIDEO PLL patches which generalized the DSS PLL code added a new
check to verify that the PLL DCO rate is not too low.

However, this breaks OMAP4 HDMI. It seems that on OMAP4, the PLL works
fine with low DCO rates, but not on OMAP5.

As the check is just a safety measure against driver bugs (if driver
calculates PLL settings badly), we can just disable the check for now.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>