OMAPDSS: panel-tlc59108: fix timings
The current timings for 1280x800 mode result in a refresh rate of ~57Hz.
Also, the pixel clock was under the minimum pclk according to the panel
spec. And, the comment in the source code does not make any sense, as it
mentions "VESA CVT 0.31M3-R" which is a 640x480 mode.

The patch fixes the timings to use the "typical" timings from the panel
spec. The exact pixel clock to get 60Hz would be 69329520, but
unfortunately it seems our video PLL (HW or SW) fails to lock to that
frequency, so the pixel clock used is 69300404. This still results very
nearly 60Hz.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>