beaglebone: added expansion header to dtb
This adds gpio and pinmux helpers to the majority of available expansion header pins
based on the cape-universal work from Charles Steinkuehler making them userspace
configurable. This is not a substitute for Capemgr as it doesn't perform the
configuration based on cape detection, nor does it enable dynamic configuration of
all types of peripherals that could be on a cape. It does, however, enable many
developers to rapidly experiment with a lesser degree of complexity.

Derived from:
https://github.com/cdsteinkuehler/beaglebone-universal-io/blob/52461b52ef3203e648399c16c7e160c848a04b5c$

Signed-off-by: Jason Kridner <jdk@ti.com>
Cc: Charles Steinkuehler <charles@steinkuehler.net>