OMAPDSS: panel-tlc59108: fix init seq for 10" display board
The same driver handles both 7" and 10" display boards, but they have
other differences than the panel size. One such difference seems to be
the configuration of the TLC59108 LED driver chip.

With the current kernel, the 10" panel has small artifacts. This seem to
be caused by wrong setting for the CLKSEL pin of the SN75LVDS83B
FlatLink chip. This pin is controlled by the LED driver chip.

This patch adds separate init sequences for the LED driver chip for 7"
and 10" boards, fixing the display artifacts.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>