mfd: ti_am335x_tscadc: fix TSC resume
In the resume path, the ADC invokes am335x_tsc_se_set_cache() with 0 as
argument if continous mode is not in use. This in turn disables the TSC
steps and so it is not working until one ADC sampling is performed.

This patch fixes it by writing the current cached mask instead of the
passed steps.

Fixes: 7ca6740cd1cd ("mfd: input: iio: ti_amm335x: Rework TSC/ADCA
synchronization")
Cc: stable@vger.kernel.org # v3.13+
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>