HACK: ARM: OMAP2+: pdata-quirks: Add am437x-gp-evm NAND/eMMC selection
Some boards e.g. am437x-gp-evm, need a GPIO to enable the
right buffer for either NAND or eMMC operation. This falls out
of scope of both the drivers and hence must be done at the board
driver level.

Add the relevant quirk for am437x-gp-evm. The GPIO will
be configured depending on presence/abscense of gpmc node.

Signed-off-by: Roger Quadros <rogerq@ti.com>