ARM: dts: dra7-evm: remove mcspi1 pins
HDMI HPD pin can be muxed to be used for mcspi1. The current
dra7-evm.dts file contains muxing for both HDMI and MCSPI, which gives
an error.

Remove the mcspi1 pinmuxing and mcspi1 node, as it is not used on
dra7-evm.

Fixes 6e58b8f (ARM: dts: DRA7: Add the dts files
        for dra7 SoC and dra7-evm board)

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Acked-by: Felipe Balbi <balbi@ti.com>