usb: dwc3: core: only reset res->start in case of error
[ Upstream commit 28f17ad6a24700b7b2594988740cee0347028713 ]

That trick is only needed if we end up with an error, so
there's no point in messing that outside of an error path.

In fact doing so causes problems when removing dwc3.ko,
problems which commit c5a1fbc (usb: dwc3: dwc3-omap: Fix
the crash on module removal) mistakenly tried to fix.

Signed-off-by: Felipe Balbi <balbi@ti.com>
Tested-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>