OMAPDSS: panel-tlc59108: fix enable-gpio polarity
The current panel-tlc59108 driver doesn't respect the ACTIVE_HIGH/LOW
setting for the enable-gpio from the .dts file.

This patch adds support for setting the polarity by using the new GPIO
API which handles the polarity automatically.

The polarity in the dra7*.dts files is also fixed, as it was wrong. The
enable-gpio needs to be set to low when enabling the display, so we need
to define the gpio as GPIO_ACTIVE_LOW.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>