OMAPDSS: fix sii9022 module file name
At the moment the makefiles create encoder_sii9022_full.ko, which is
somewhat odd name for the module. It should be encoder-sii9022.ko, like
all the other modules.

This patch renames encoder-sii9022.c to encoder-sii9022-video.c so that
we can create encoder-sii9022.ko from the video and audio parts.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>