iommu/omap: revise check for valid arch_data in attach_dev
Commit a894073 (iommu/omap: check for valid archdata in attach_dev)
has added a sanity check to make sure that the archdata is valid in
attach_dev. Revise this check to also check for the name field, which
would be set correctly only fir properly configured client devices.
The check is also moved outside of the locking primitive.

This revision syncs up this check with the code in upstream vanilla
kernel.

Signed-off-by: Suman Anna <s-anna@ti.com>