drm/omap: add support for data/sync edges and DE polarity
DRM uses struct drm_display_mode to pass video mode timings, and as it
comes from the desktop world, it's missing flags to represent when data/sync
should be driven or if DE should be active high or low.

The support for those should be added to the DRM core, but for the time
being let's just add custom flags for omapdrm.

This fixes the signal polarities for panels that use different settings
than what omapdrm uses.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>