drm/tilcdc: make frame completion interrupt always enabled
DRM allows vblank handling to be controlled dynamically.  Therefore if vblank
interrupts are not needed by anybody they are disabled.  However on tilcdc a
ping-pong dma is used and there is no way to find out from the hardware which
dma channel will be used for the next frame to program the corresponding dma
registers with the new framebuffer address. The only way currently known is to
use the vsync handler to track the channel currently used.  This means that
disabling the vsync is a bad idea because it is always needed to track the dma
channel.

This patch removes the enable/disable_vblank functions, moves the enabling of
the FRAME0/FRAME1_DONE interrupts into the post_install_interrupt handler so
they are always active and removes the calls to drm_vblank_get,
drm_vblank_put.

Signed-off-by: Darren Etheridge <detheridge@ti.com>