ARM/dmaengine: edma: Move dmaengine pdevice registration to arm/common/edma.c
The platform device for the eDMA dmaengine driver should be created in the
arch/arm/common/edma.c driver and not by itself. This way we can avoid to
have extra (unconnected) logic in the edma dmaengine driver to create
correct number of devices (one or two) and also this way we can make sure
that the platform device for the dmaengine part is only registered when
the device is using edma. Currently the first device is registered without
unconditionally, and the dmaengine driver will fail since the platform does
not have support for eDMA.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>