ARM: dts: am335x-evmsk: enable second USB controller
The first port is hard wired to a USB-HUB and should be operated in
device mode (or OTG). The second port (which is enabled by this patch)
is wired to a host port. However the ID-pin isn't so we have to force
host mode here.
There is no match thing upstream commit for this because upstream has
already both ports enabled with USB1 forced in host mode.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
[nsekhar@ti.com: made it explicit that USB1 is forced in host
                 mode even in upstream]
Signed-off-by: Sekhar Nori <nsekhar@ti.com>