arm: mach-omap2: omap_hwmod_33xx_43xx_ipblock_data: Move am33xx_l4_hs_hwmod to ipblock_data
This patch moves am33xx_l4_hs_hwmod from 33xx_data to ipblock_data. This
is required because it can be reused in 43xx to add hwmod support to
tscadc as it uses the same l4_hs hwmod in am437x. This is helps in setting
appropriate clock domain names for tscadc on am33xx and am43xx.

Signed-off-by: Vignesh R <vigneshr@ti.com>