ARM: dts: am335x-evm: Fix backlight sleep pin config
On the newest AM335x GP EVM boards, backlight remains on during suspend.
According to "General Purpose Daughterboard SKU1 Modifications" document
version 1.05, CPLD code has been modified to add an inverter for
LCD_ENBKL signal so the EVM does not have to keep a GPIO high during
suspend to power off the backlight. Solution for this is to use a pulldown
during suspend to make sure the signal is low and the backlight gets
powered off.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>