net/rpmsg: store remote processor id in rpmsg_socket
The rpmsg_proto depends on rpmsg_sock_get_proc_id() to retrieve
the remote processor address at runtime, but this relies on
retrieving the address based on looking up the underlying parent
virtio device of the associated rpmsg channel. This can not be
always relied on, as the virtio devices can be deleted and
recreated during a recovery of a remote processor.

Add a new field 'rproc_id' to the rpmsg_socket structure, and
use this field to store the remote processor address during
the connect and bind socket calls. The rpmsg sockets leverage
connect for Tx sockets and bind for Rx sockets to exchange
messages to/from a remote processor. This field allows runtime
lookup of the processor id in any of the operations on the
socket, without being burdened with the life cycle of the
underlying rpmsg or virtio devices.

Signed-off-by: Suman Anna <s-anna@ti.com>