rpmsg: rpc: return proper error status during recovery in read
An user thread can be blocked while waiting for new response
messages. The thread is unblocked either when a new message is
available for the thread to read or when the remote processor
has crashed and the rpmsg-rpc device has been removed and
recreated as part of the recovery process.

A special error number ENXIO is returned in the file operations
on the file descriptors for the rpmsg-rpc driver to notify the
user process that the underlying device is invalid due to a
crash of the remote processor. An incorrect error number of
EIO is being returned currently if the user thread blocked on
a read is woken up due to a recovery, fix the same to return
the correct value of -ENXIO so that the user process can take
an appropriate action.

Signed-off-by: Suman Anna <s-anna@ti.com>